refactor structure project
