--- src/buyers/controllers/buyers.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { BuyersController } from './buyers.controller';
-
-describe('BuyersController', () => {
-  let controller: BuyersController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [BuyersController],
-    }).compile();
-
-    controller = module.get<BuyersController>(BuyersController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
