--- src/categories/categories.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { CategoriesService } from './services/categories.service';
-import { CategoriesController } from './controllers/categories.controller';
-
-@Module({
-  providers: [CategoriesService],
-  controllers: [CategoriesController],
-})
-export class CategoriesModule {}
