--- src/categories/entities/Category.entity.ts
@@ -1,5 +1,0 @@
-export class Category {
-  id: number;
-  name: string;
-  description: string;
-}
