--- src/manufacturers/manufacturers.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ManufacturersController } from './controllers/manufacturers.controller';
-import { ManufacturersService } from './services/manufacturers.service';
-
-@Module({
-  controllers: [ManufacturersController],
-  providers: [ManufacturersService],
-})
-export class UsersModule {}
