+++ src/operators/controllers/buyers.controller.spec.ts
@@ -1,0 +1,18 @@
+// import { Test, TestingModule } from '@nestjs/testing';
+// import { BuyersController } from '../../buyers/controllers/buyers.controller';
+
+// describe('BuyersController', () => {
+//   let controller: BuyersController;
+
+//   beforeEach(async () => {
+//     const module: TestingModule = await Test.createTestingModule({
+//       controllers: [BuyersController],
+//     }).compile();
+
+//     controller = module.get<BuyersController>(BuyersController);
+//   });
+
+//   it('should be defined', () => {
+//     expect(controller).toBeDefined();
+//   });
+// });
