+++ src/operators/controllers/orders.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { OrdersController } from '../../orders/controllers/orders.controller';
+import { OrdersController } from '../controllers/orders.controller';
 
 describe('OrdersController', () => {
   let controller: OrdersController;
