+++ src/operators/services/orders.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { OrdersService } from '../../orders/services/orders.service';
+import { OrdersService } from '../services/orders.service';
 
 describe('OrdersService', () => {
   let service: OrdersService;
