+++ src/products/products.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ProductsService } from './services/products.service';
 import { ProductsController } from './controllers/products.controller';
-import { CategoriesService } from 'src/categories/services/categories.service';
-import { ManufacturersService } from 'src/manufacturers/services/manufacturers.service';
-import { ManufacturersController } from 'src/manufacturers/controllers/manufacturers.controller';
-import { CategoriesController } from 'src/categories/controllers/categories.controller';
+import { CategoriesService } from './services/categories.service';
+import { ManufacturersService } from 'src/products/services/manufacturers.service';
+import { ManufacturersController } from 'src/products/controllers/manufacturers.controller';
+import { CategoriesController } from 'src/products/controllers/categories.controller';
 
 @Module({
   controllers: [
